refactor(cmd): split main.go into main, free, tuner, and serial files
