+++ cmd/ranan/tuner.go
@@ -1,0 +1,104 @@
+package main
+
+import (
+	"bufio"
+	"fmt"
+	"log"
+	"os"
+	"strings"
+
+	"github.com/kaidi-ahas/ranan/internal/audio"
+	"github.com/kaidi-ahas/ranan/internal/music"
+	"github.com/kaidi-ahas/ranan/internal/pitch"
+	"github.com/kaidi-ahas/ranan/internal/serial"
+)
+
+func RunTunerMode(mic *audio.Microphone, buf *pitch.Buffer, port *serial.Port, stop chan os.Signal, threshold float64) {
+	inputCh := make(chan string)
+
+	readInput := func() {
+		scanner := bufio.NewScanner(os.Stdin)
+		for scanner.Scan() {
+			inputCh <- scanner.Text()
+		}
+		close(inputCh)
+	}
+
+	go readInput()
+
+	promptNext := make(chan struct{}, 1)
+	promptNext <- struct{}{}
+
+	for {
+		select {
+		case <-stop:
+			fmt.Println("\nStopped.")
+			return
+		case <-promptNext:
+			fmt.Print("Enter target note (e.g. A4) or type 'q' to quit: ")
+			input, ok := <-inputCh
+			if !ok {
+				return
+			}
+
+			if strings.TrimSpace(input) == "q" {
+				fmt.Println("Stopped.")
+				return
+			}
+
+			input = strings.TrimSpace(input)
+			name, octave, err := music.ParseNote(input)
+			if err != nil {
+				fmt.Printf("Invalid note: %v\n", err)
+				promptNext <- struct{}{}
+				continue
+			}
+
+			targetFreq, err := music.ToFrequency(name, octave)
+			if err != nil {
+				fmt.Printf("Invalid note: %v\n", err)
+				promptNext <- struct{}{}
+				continue
+			}
+
+			targetLabel := fmt.Sprintf("%s%d", name, octave)
+			fmt.Printf("Tuning to %s (%.2f Hz)...\n", targetLabel, targetFreq)
+
+			go func(label string, freq float64) {
+				for {
+					select {
+					case <-stop:
+						return
+					default:
+						note, ok := ReadNote(mic, buf)
+						if !ok {
+							continue
+						}
+
+						centsFromTarget := music.CentsBetween(note.Frequency, freq)
+
+						fmt.Printf("\rTarget: %s | Frequency: %.2f Hz | Cents: %+.2f   ",
+							label,
+							note.Frequency,
+							centsFromTarget,
+						)
+
+						if port != nil {
+							status := music.TuningStatus(centsFromTarget, threshold)
+							if err := port.Send(note.Name, note.Octave, status); err != nil {
+								log.Printf("failed to send to Arduino: %v", err)
+							}
+						}
+
+						if centsFromTarget >= -threshold && centsFromTarget <= threshold {
+							fmt.Printf("\nNote %s is in tune. Press Enter to confirm and continue.", label)
+							<-inputCh
+							promptNext <- struct{}{}
+							return
+						}
+					}
+				}
+			}(targetLabel, targetFreq)
+		}
+	}
+}
